requirements.txt
django==2.0
gunicorn==19.9.0
whitenoise==4.1.2
asgiref==3.2.7
cycler==0.10.0
django-cors-headers==3.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
hmmlearn==0.2.3
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
psycopg2-binary==2.8.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.0.0
